requirements.txt
aiohttp==3.6.2
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
billiard==3.3.0.23
blinker==1.4
boto==2.49.0
brainyquote==0.2
cachetools==4.1.0
celery==3.1.26.post2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
discord.py==1.3.3
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-background-tasks==1.2.5
django-celery==3.3.1
django-compat==1.0.15
django-crontab==0.7.1
django-extensions==2.2.9
django-storages==1.9.1
django-toolbelt==0.0.1
dnspython==1.16.0
email-validator==1.0.5
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
geojson==2.5.0
google-auth==1.17.2
google-auth-oauthlib==0.4.1
gspread==3.5.0
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==3.0.37
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==7.1.2
pipenv==2018.11.26
praw==7.0.0
prawcore==1.3.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.4.4
pyowm==3.0.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytube==9.6.0
pytube3==9.6.4
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
static3==0.7.0
tabulate==0.8.7
terminaltables==3.1.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzlocal==2.1
update-checker==0.17
urllib3==1.25.9
virtualenv==20.0.18
virtualenv-clone==0.5.4
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.1.0
wrapt==1.11.2
WTForms==2.3.1
yarl==1.4.2
zipp==3.1.0
